alertas-function/requirements.txt
azure-functions
pymssql
requests